Adds fitting_and_plotting.py and saving.py, which contain the fitting, plotting, and saving    functions that used to be methods in the BaseDataAnalysis. These are now removed from    BaseDataAnalysis. Keeps only calls to extract_data and process_data in BaseDataAnalysis. Changes the data_filter variable in filter from a type function to a type string that will    be evaluated. Adds get_cal_state_color, get_latex_prob_label, prepare_1d_plot to data_processing.py. Adds calls to prepare_fitting, run_fitting, analyze_fit_results, prepare_plot, plot to RabiAnalysis. Adds __call__ method to RabiAnalysis. Adds call to save_fit_results in run_fitting. Adds call to save_figures in plot.
